fix: Use POST method in localhost and use GET method in production for static image api
