+++ sites/geohub/src/routes/api/style/[id]/static/[bbox]/[width]x[height].[format]/+server.ts
@@ -1,16 +1,27 @@
 import type { RequestHandler } from './$types';
 import { error } from '@sveltejs/kit';
-import { env } from '$env/dynamic/private';
 import { getStyleById } from '$lib/server/helpers';
 import type { DashboardMapStyle } from '$lib/types';
 import { getDomainFromEmail } from '$lib/helper';
 import { AccessLevel } from '$lib/config/AppConfig';
+import { env } from '$env/dynamic/private';
 
-export const GET: RequestHandler = async ({ locals, params, url, fetch }) => {
+export const GET: RequestHandler = async ({ locals, params, url }) => {
 	const session = await locals.getSession();
 
+	const bbox = params.bbox;
 	const width = Number(params.width);
 	const height = Number(params.height);
+
+	const ratio = url.searchParams.get('ratio') ? Number(url.searchParams.get('ratio')) : 1;
+	if (!(ratio === 1 || ratio === 2)) {
+		throw error(400, 'ratio should be either 1 or 2.');
+	}
+
+	const format = params.format;
+	if (!['jpeg', 'png', 'webp'].includes(format)) {
+		throw error(400, 'Unsupported format.');
+	}
 
 	const id = params.id;
 	const style = (await getStyleById(
@@ -37,7 +48,7 @@ export const GET: RequestHandler = async ({ locals, params, url, fetch }) => {
 		}
 	}
 
-	const apiUrl = `${env.GEOHUB_STATIC_IMAGE_API}/style/static/auto/${width}x${height}.png`;
+	const apiUrl = `${env.GEOHUB_STATIC_IMAGE_API}/style/static/${bbox}/${width}x${height}.${format}?ratio=${ratio}`;
 	const res = await fetch(apiUrl, {
 		method: 'POST',
 		body: JSON.stringify(style.style)
@@ -49,7 +60,7 @@ export const GET: RequestHandler = async ({ locals, params, url, fetch }) => {
 
 	return new Response(image, {
 		headers: {
-			'Content-type': 'image/png'
+			'Content-type': `image/${format}`
 		}
 	});
 };
