+++ sites/geohub/src/routes/api/style/[id]/static/auto/[width]x[height].[format]/+server.ts
@@ -1,21 +1,26 @@
 import type { RequestHandler } from './$types';
 import { error } from '@sveltejs/kit';
+import { env } from '$env/dynamic/private';
 import { getStyleById } from '$lib/server/helpers';
 import type { DashboardMapStyle } from '$lib/types';
 import { getDomainFromEmail } from '$lib/helper';
 import { AccessLevel } from '$lib/config/AppConfig';
-import { env } from '$env/dynamic/private';
 
-export const GET: RequestHandler = async ({ locals, params, url }) => {
+export const GET: RequestHandler = async ({ locals, params, url, fetch }) => {
 	const session = await locals.getSession();
 
-	const lon = Number(params.lon);
-	const lat = Number(params.lat);
-	const zoom = Number(params.zoom);
-	const bearing = Number(params.bearing);
-	const pitch = Number(params.pitch);
 	const width = Number(params.width);
 	const height = Number(params.height);
+
+	const ratio = url.searchParams.get('ratio') ? Number(url.searchParams.get('ratio')) : 1;
+	if (!(ratio === 1 || ratio === 2)) {
+		throw error(400, 'ratio should be either 1 or 2.');
+	}
+
+	const format = params.format;
+	if (!['jpeg', 'png', 'webp'].includes(format)) {
+		throw error(400, 'Unsupported format.');
+	}
 
 	const id = params.id;
 	const style = (await getStyleById(
@@ -42,7 +47,7 @@ export const GET: RequestHandler = async ({ locals, params, url }) => {
 		}
 	}
 
-	const apiUrl = `${env.GEOHUB_STATIC_IMAGE_API}/style/static/${lon},${lat},${zoom},${bearing},${pitch}/${width}x${height}.png`;
+	const apiUrl = `${env.GEOHUB_STATIC_IMAGE_API}/style/static/auto/${width}x${height}.${format}?ratio=${ratio}`;
 	const res = await fetch(apiUrl, {
 		method: 'POST',
 		body: JSON.stringify(style.style)
@@ -54,7 +59,7 @@ export const GET: RequestHandler = async ({ locals, params, url }) => {
 
 	return new Response(image, {
 		headers: {
-			'Content-type': 'image/png'
+			'Content-type': `image/${format}`
 		}
 	});
 };
